çalışan üzerinde kayıtlı olan vardiyaların silinme işlemini engelleyen olasılıkların hata mesajları eklendi silinmesi engellendi.

+++ BilgeHotelProject/WebUI/Areas/HumanResources/Controllers/ShiftController.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using WebUI.Models.Shift;
 using Core.Entities.Enum;
 using Core.Utilities.Results.Abstract;
+using Entities.Concrete;
 
 namespace WebUI.Areas.HumanResources.Controllers
 {
@@ -19,12 +20,14 @@ namespace WebUI.Areas.HumanResources.Controllers
         private readonly IMapper mapper;
         private readonly IShiftService shiftService;
         private readonly IResult result;
+        private readonly IEmployeeService employeeService;
 
-        public ShiftController(IMapper mapper, IShiftService shiftService, IResult result)
+        public ShiftController(IMapper mapper, IShiftService shiftService, IResult result, IEmployeeService employeeService)
         {
             this.mapper = mapper;
             this.shiftService = shiftService;
             this.result = result;
+            this.employeeService = employeeService;
         }    
         public async Task<IActionResult> Index()
         {
@@ -42,16 +45,25 @@ namespace WebUI.Areas.HumanResources.Controllers
             var shift = await shiftService.GetById(id);
             if (shift!=null)
             {
-                if (shift.Status==Status.Active)
+                if (shift.Status==Status.Active && (await employeeService.Any(x => x.ShiftID == id)) == false)
                 {
                     shift.Status = Status.Deleted;
+                    var changeResult = shiftService.Update(shift);
+                    TempData["ShiftResult"] = JsonConvert.SerializeObject(changeResult);
+                }
+                else if (shift.Status == Status.Deleted)
+                {
+                    shift.Status = Status.Active;
+                    var changeResult = shiftService.Update(shift);
+                    TempData["ShiftResult"] = JsonConvert.SerializeObject(changeResult);
                 }
                 else
                 {
-                    shift.Status = Status.Active;
+                    result.ResultStatus = ResultStatus.Error;
+                    result.Message = "Silmek istediğiniz vardiya bir veya daha fazla çalışan üzerinde kayıtlı olabilir. Lütfen çalışan vardiyalarını güncelleyin.";
+                    TempData["ShiftResult"] = JsonConvert.SerializeObject(result);
                 }
-                var changeResult = shiftService.Update(shift);
-                TempData["ShiftResult"] = JsonConvert.SerializeObject(changeResult);
+                
             }
             else
             {
@@ -64,7 +76,7 @@ namespace WebUI.Areas.HumanResources.Controllers
         public async Task<IActionResult> RemoveForce(int id)
         {
             
-            if (await shiftService.Any(x=>x.ID==id))
+            if (await shiftService.Any(x=>x.ID==id) && (await employeeService.Any(x=>x.ShiftID==id))==false)
             {
                 var changeResult = shiftService.RemoveForce(id);
                 TempData["ShiftResult"] = JsonConvert.SerializeObject(changeResult);
@@ -72,10 +84,36 @@ namespace WebUI.Areas.HumanResources.Controllers
             else
             {
                 result.ResultStatus = ResultStatus.Error;
-                result.Message = "İlgili idye ait vardiya bulunamadı.";
+                result.Message = "İlgili idye ait vardiya bulunamadı. Ya da silmek istediğiniz vardiya bir veya daha fazla çalışan üzerinde kayıtlı olabilir.";
                 TempData["ShiftResult"] = JsonConvert.SerializeObject(result);
             }
             return RedirectToAction("Index");
+        }
+        public IActionResult CreateShift()
+        {
+            if (TempData["ShiftResult"] != null)
+            {
+                var shiftResultResult = JsonConvert.DeserializeObject<Result>(TempData["ShiftResult"].ToString());
+                ViewBag.ShiftResult = shiftResultResult;
+            }
+            return View();
+        }
+        [HttpPost]
+        public IActionResult CreateShift(VMShiftCreate vMShiftCreate)
+        {
+            if (ModelState.IsValid)
+            {
+                var shift = mapper.Map<Shift>(vMShiftCreate);
+                var createResult = shiftService.Create(shift);
+                TempData["ShiftResult"] = JsonConvert.SerializeObject(createResult);
+
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return View(vMShiftCreate);
+            }
+            
         }
     }
 }
